+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add `.clang-format` draft
 - Add `lwow_match_or_skip_rom`
 - Add `lwow_ds18x20_get_alarm_temp` and `lwow_ds18x20_get_temp_conversion_time`
+- Remove deprecated functions, prepare for version `4.0.0`
 
 ## v3.0.2